Componentization & fix color version
